generate monaco entry point using a real hash.
The previous hash was based on its content, not transitively on its
inputs. If monaco was updated but the dependencies weren't, the hash
would stay the same. Instead, let EsBuild generate the hash. We can get
the file name through its position in the metafile.

Note that css and javascript have different hashes.